Comments and splitting out ab-utils.js
