+++ static/src/javascripts/projects/common/modules/experiments/ab-local-storage.js
@@ -1,82 +1,13 @@
 // @flow
 import { local } from 'lib/storage';
-import {
-    NOT_IN_TEST,
-    notInTestVariant,
-    participationsKey,
-} from 'common/modules/experiments/ab-constants';
+import { NOT_IN_TEST, participationsKey, } from './ab-constants';
+import { filterParticipations, testAndParticipationsToVariant } from './ab-utils';
 
-export const participationsToArray = (
-    participations: Participations
-): { testId: string, variantId: string }[] =>
-    Object.keys(participations).map(testId => ({
-        testId,
-        variantId: participations[testId].variant,
-    }));
-
-export const arrayToParticipations = (
-    arr: { testId: string, variantId: string }[]
-): Participations =>
-    arr.reduce(
-        (obj, { testId, variantId }) => ({
-            ...obj,
-            [testId]: { variant: variantId },
-        }),
-        {}
-    );
-
-export const filterParticipations = (
-    participations: Participations,
-    filter: ({ testId: string, variantId: string }) => boolean
-): Participations =>
-    arrayToParticipations(participationsToArray(participations).filter(filter));
-
+// -------
+// Reading
+// -------
 export const getParticipationsFromLocalStorage = (): Participations =>
     local.get(participationsKey) || {};
-
-export const setParticipationsInLocalStorage = (
-    participations: Participations
-): void => {
-    local.set(participationsKey, participations);
-};
-
-// If the given test has a variant which is specified in the participations, return it.
-// Also return a 'notintest' variant, if such a participation is present for the supplied test.
-export const variantFromParticipations = (
-    test: ABTest,
-    participations: Participations
-): ?Variant => {
-    const participation = participations[test.id];
-    if (participation) {
-        // We need to return something concrete here to ensure
-        // that a notintest variant actually prevents other variants running.
-        if (participation.variant === NOT_IN_TEST) {
-            return notInTestVariant;
-        }
-
-        return test.variants.find(
-            variant => variant.id === participation.variant
-        );
-    }
-
-    return null;
-};
-
-export const runnableTestsToParticipations = (
-    runnableTests: $ReadOnlyArray<Runnable<ABTest>>
-): Participations => {
-    const participations: Participations = {};
-    runnableTests.forEach(({ id: testId, variantToRun }) => {
-        participations[testId] = { variant: variantToRun.id };
-    });
-
-    return participations;
-};
-
-// Wipes all localStorage participations
-export const clearParticipations = (): void => {
-    local.remove(participationsKey);
-};
 
 export const getTestExclusionsFromLocalStorage = (): Participations =>
     filterParticipations(
@@ -86,4 +17,20 @@ export const getTestExclusionsFromLocalStorage = (): Participations =>
 
 // If the given test has a variant specified in localStorage, return it.
 export const getVariantFromLocalStorage = (test: ABTest): ?Variant =>
-    variantFromParticipations(test, getParticipationsFromLocalStorage());
+    testAndParticipationsToVariant(test, getParticipationsFromLocalStorage());
+
+
+// -------
+// Writing
+// -------
+
+// Wipes all localStorage participations
+export const clearParticipations = (): void => {
+    local.remove(participationsKey);
+};
+
+export const setParticipationsInLocalStorage = (
+    participations: Participations
+): void => {
+    local.set(participationsKey, participations);
+};
