+++ static/src/javascripts/projects/common/modules/experiments/ab-url.js
@@ -1,10 +1,7 @@
 // @flow
 
 import { NOT_IN_TEST } from 'common/modules/experiments/ab-constants';
-import {
-    filterParticipations,
-    variantFromParticipations,
-} from 'common/modules/experiments/ab-local-storage';
+import { testAndParticipationsToVariant, filterParticipations } from './ab-utils';
 
 export const getForcedParticipationsFromUrl = (): Participations => {
     if (window.location.hash.startsWith('#ab')) {
@@ -23,7 +20,7 @@ export const getForcedParticipationsFromUrl = (): Participations => {
 
 // If the given test has a variant which is forced by the URL, return it
 export const getVariantFromUrl = (test: ABTest): ?Variant =>
-    variantFromParticipations(test, getForcedParticipationsFromUrl());
+    testAndParticipationsToVariant(test, getForcedParticipationsFromUrl());
 
 export const getTestExclusionsFromUrl = (): Participations =>
     filterParticipations(
